requirements.txt
torch>=2.0.0
numpy>=1.21.0
pyyaml>=5.4.0
pypinyin>=0.44.0
g2p-en>=2.1.0
onnx>=1.12.0
onnxruntime>=1.12.0
soundfile>=0.10.0
matplotlib>=3.5.0
scipy>=1.7.0
